eeg-processor/requirements.txt
# requirements.txt for EEG processor service
numpy>=1.20.0
scipy>=1.6.0
pandas>=1.2.0
scikit-learn>=0.24.0
mne>=0.22.0
pymongo>=3.11.0
python-dotenv>=0.15.0
pybids>=0.13.1
watchdog>=2.0.0
joblib>=1.0.0
matplotlib>=3.3.0